feat: add bubble menu to mobile toolbar
