+++ .changeset/calm-kings-cheer.md
@@ -1,0 +1,5 @@
+---
+'@sigle/app': minor
+---
+
+Add bubble menu to mobile toolbar.
