+++ sigle/src/modules/editor/BubbleMenu.tsx
@@ -1,17 +1,8 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import { Editor, BubbleMenu as TipTapBubbleMenu } from '@tiptap/react';
 import { isTextSelection } from '@tiptap/core';
-import {
-  Link1Icon,
-  FontBoldIcon,
-  FontItalicIcon,
-  CodeIcon,
-  Cross2Icon,
-  StrikethroughIcon,
-  UnderlineIcon,
-} from '@radix-ui/react-icons';
 import { globalCss, styled } from '../../stitches.config';
-import { Flex } from '../../ui';
+import { BubbleMenuItems } from './BubbleMenuItems';
 
 // Tippyjs theme used by the bubble menu
 const globalStylesCustomEditor = globalCss({
@@ -20,19 +11,35 @@ const globalStylesCustomEditor = globalCss({
   },
   ".tippy-box[data-theme~='sigle-editor-bubble-menu'][data-placement^='top'] > .tippy-arrow::before":
     {
+      display: 'none',
       borderTopColor: '$gray11',
+      '@xl': {
+        display: 'block',
+      },
     },
   ".tippy-box[data-theme~='sigle-editor-bubble-menu'][data-placement^='bottom'] > .tippy-arrow::before":
     {
+      display: 'none',
       borderBottomColor: '$gray11',
+      '@xl': {
+        display: 'block',
+      },
     },
   ".tippy-box[data-theme~='sigle-editor-bubble-menu'][data-placement^='left'] > .tippy-arrow::before":
     {
+      display: 'none',
       borderLeftColor: '$gray11',
+      '@xl': {
+        display: 'block',
+      },
     },
   ".tippy-box[data-theme~='sigle-editor-bubble-menu'][data-placement^='right'] > .tippy-arrow::before":
     {
+      display: 'none',
       borderRightColor: '$gray11',
+      '@xl': {
+        display: 'block',
+      },
     },
   ".tippy-box[data-theme~='sigle-editor-bubble-menu'] .tippy-content": {
     padding: 0,
@@ -41,109 +48,30 @@ const globalStylesCustomEditor = globalCss({
 });
 
 const StyledBubbleMenu = styled(TipTapBubbleMenu, {
-  display: 'flex',
+  display: 'none',
   gap: '$3',
   br: '$1',
   py: '$3',
   px: '$4',
-});
 
-const BubbleMenuButton = styled('button', {
-  color: '$gray1',
-
-  variants: {
-    active: {
-      true: {
-        color: '$orange10',
-      },
-    },
+  '@xl': {
+    display: 'flex',
   },
-});
-
-const BubbleMenuInput = styled('input', {
-  width: '100%',
-  pl: '$2',
-  pr: '$1',
-  backgroundColor: 'transparent',
-  outline: 'none',
 });
 
 // Maybe can be used on clicks https://github.com/ueberdosis/tiptap/issues/104#issuecomment-912794709
 
 interface BubbleMenuProps {
   editor: Editor;
+  defaultOpen?: boolean;
 }
 
-export const BubbleMenu = ({ editor }: BubbleMenuProps) => {
+export const BubbleMenu = ({
+  editor,
+  defaultOpen = false,
+}: BubbleMenuProps) => {
   globalStylesCustomEditor();
-
   const [linkState, setLinkState] = useState({ open: false, value: '' });
-
-  // Listen to any key press to detect cmd + k and activate the link edition
-  useEffect(() => {
-    function onKeyDown(event: KeyboardEvent) {
-      // We want all our commands to start with the user pressing ctrl or cmd for mac users
-      if ((event.ctrlKey || event.metaKey) && event.key === 'k') {
-        onSelectLink();
-      }
-    }
-
-    window.addEventListener('keydown', onKeyDown);
-    return () => window.removeEventListener('keydown', onKeyDown);
-  }, []);
-
-  const onSelectLink = () => {
-    // Get href of selected link to pre fill the input
-    const existingHref = editor.isActive('link')
-      ? editor.getAttributes('link').href
-      : '';
-
-    setLinkState({
-      open: true,
-      value: existingHref,
-    });
-  };
-
-  const onSubmitLink = (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    let safeLinkValue = linkState.value.trim();
-
-    if (
-      safeLinkValue &&
-      !safeLinkValue.startsWith('http') &&
-      !safeLinkValue.startsWith('#')
-    ) {
-      safeLinkValue = `https://${linkState.value}`;
-    }
-
-    if (safeLinkValue) {
-      editor
-        .chain()
-        .focus()
-        .extendMarkRange('link')
-        .setLink({ href: safeLinkValue })
-        // Set the text selection at the end of the link selection
-        // that way user can continue to type easily
-        .setTextSelection(editor.state.selection.$to.pos)
-        // Unset link selection se when the user continues to type it won't be a link
-        .unsetLink()
-        .run();
-    } else {
-      // If input text is empty we unset the link
-      editor.chain().focus().extendMarkRange('link').unsetLink().run();
-    }
-
-    resetLink();
-  };
-
-  const onKeyDown = (event: React.KeyboardEvent) => {
-    // If user press escape we hide the link input
-    if (event.key === 'Escape') {
-      event.preventDefault();
-      resetLink();
-    }
-  };
 
   const resetLink = () => {
     setLinkState({
@@ -160,6 +88,9 @@ export const BubbleMenu = ({ editor }: BubbleMenuProps) => {
         onHidden: () => resetLink(),
       }}
       shouldShow={({ editor, state, from, to, view }) => {
+        if (defaultOpen) {
+          return true;
+        }
         // Take the initial implementation of the plugin and extends it
         // https://github.com/ueberdosis/tiptap/blob/main/packages/extension-bubble-menu/src/bubble-menu-plugin.ts#L43
         const { doc, selection } = state;
@@ -194,65 +125,7 @@ export const BubbleMenu = ({ editor }: BubbleMenuProps) => {
       }}
       editor={editor}
     >
-      {!linkState.open ? (
-        <>
-          <BubbleMenuButton
-            onClick={() => editor.chain().focus().toggleBold().run()}
-            active={editor.isActive('bold')}
-          >
-            <FontBoldIcon height={18} width={18} />
-          </BubbleMenuButton>
-          <BubbleMenuButton
-            onClick={() => editor.chain().focus().toggleItalic().run()}
-            active={editor.isActive('italic')}
-          >
-            <FontItalicIcon height={18} width={18} />
-          </BubbleMenuButton>
-          <BubbleMenuButton
-            onClick={() => editor.chain().focus().toggleUnderline().run()}
-            active={editor.isActive('underline')}
-          >
-            <UnderlineIcon height={18} width={18} />
-          </BubbleMenuButton>
-          <BubbleMenuButton
-            onClick={() => editor.chain().focus().toggleStrike().run()}
-            active={editor.isActive('strike')}
-          >
-            <StrikethroughIcon height={18} width={18} />
-          </BubbleMenuButton>
-          <BubbleMenuButton
-            onClick={() => editor.chain().focus().toggleCode().run()}
-            active={editor.isActive('code')}
-          >
-            <CodeIcon height={18} width={18} />
-          </BubbleMenuButton>
-          <BubbleMenuButton
-            onClick={() => onSelectLink()}
-            active={editor.isActive('link')}
-          >
-            <Link1Icon height={18} width={18} />
-          </BubbleMenuButton>
-        </>
-      ) : (
-        <Flex as="form" onSubmit={onSubmitLink}>
-          <BubbleMenuInput
-            value={linkState.value}
-            onKeyDown={onKeyDown}
-            onChange={(e) =>
-              setLinkState((state) => ({ ...state, value: e.target.value }))
-            }
-            placeholder="Enter link ..."
-            autoFocus
-          />
-          <BubbleMenuButton
-            type="button"
-            onClick={() => resetLink()}
-            active={false}
-          >
-            <Cross2Icon height={18} width={18} />
-          </BubbleMenuButton>
-        </Flex>
-      )}
+      <BubbleMenuItems editor={editor} />
     </StyledBubbleMenu>
   );
 };
